+++ src/Contracts/Toolkit/Loggable/Contextual.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * Copyright 2025 Cloud Creativity Limited
+ *
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+declare(strict_types=1);
+
+namespace CloudCreativity\Modules\Contracts\Toolkit\Loggable;
+
+interface Contextual
+{
+    /**
+     * Get the value to use when adding the value to log context.
+     *
+     * @return mixed
+     */
+    public function context(): mixed;
+}