+++ tests/Unit/Application/Bus/Middleware/LogMessageDispatchTest.php
@@ -13,11 +13,10 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Tests\Unit\Application\Bus\Middleware;
 
 use CloudCreativity\Modules\Application\Bus\Middleware\LogMessageDispatch;
+use CloudCreativity\Modules\Contracts\Toolkit\Loggable\ContextFactory;
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\Command;
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\Message;
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\Query;
-use CloudCreativity\Modules\Toolkit\Loggable\ObjectContext;
-use CloudCreativity\Modules\Toolkit\Loggable\ResultContext;
 use CloudCreativity\Modules\Toolkit\Result\Result;
 use LogicException;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -30,7 +29,12 @@ class LogMessageDispatchTest extends TestCase
     /**
      * @var LoggerInterface&MockObject
      */
-    private LoggerInterface $logger;
+    private LoggerInterface&MockObject $logger;
+
+    /**
+     * @var MockObject&ContextFactory
+     */
+    private ContextFactory&MockObject $context;
 
     /**
      * @var Command
@@ -45,10 +49,17 @@ class LogMessageDispatchTest extends TestCase
         parent::setUp();
 
         $this->logger = $this->createMock(LoggerInterface::class);
-        $this->message = new class () implements Command {
-            public string $foo = 'bar';
-            public string $baz = 'bat';
-        };
+        $this->context = $this->createMock(ContextFactory::class);
+        $this->message = new class () implements Command {};
+    }
+
+    /**
+     * @return void
+     */
+    protected function tearDown(): void
+    {
+        parent::tearDown();
+        unset($this->logger, $this->context, $this->message);
     }
 
     /**
@@ -68,7 +79,19 @@ class LogMessageDispatchTest extends TestCase
                 return true;
             });
 
-        $middleware = new LogMessageDispatch($this->logger);
+        $context1 = ['message' => 'blah!'];
+        $context2 = ['result' => 'blah!'];
+
+        $this->context
+            ->expects($this->exactly(2))
+            ->method('make')
+            ->willReturnCallback(fn (object $in) => match ($in) {
+                $this->message => $context1,
+                $expected => $context2,
+                default => $this->fail('Unexpected object to convert to context.'),
+            });
+
+        $middleware = new LogMessageDispatch($this->logger, context: $this->context);
         $actual = $middleware($this->message, function (Message $received) use ($expected) {
             $this->assertSame($this->message, $received);
             return $expected;
@@ -76,8 +99,8 @@ class LogMessageDispatchTest extends TestCase
 
         $this->assertSame($expected, $actual);
         $this->assertSame([
-            [LogLevel::DEBUG, "Bus dispatching {$name}.", ObjectContext::from($this->message)->context()],
-            [LogLevel::INFO, "Bus dispatched {$name}.", ResultContext::from($expected)->context()],
+            [LogLevel::DEBUG, "Bus dispatching {$name}.", ['command' => $context1]],
+            [LogLevel::INFO, "Bus dispatched {$name}.", ['result' => $context2]],
         ], $logs);
     }
 
@@ -98,7 +121,20 @@ class LogMessageDispatchTest extends TestCase
                 return true;
             });
 
-        $middleware = new LogMessageDispatch($this->logger, LogLevel::NOTICE, LogLevel::WARNING);
+        $context1 = ['message' => 'blah!'];
+        $context2 = ['result' => 'blah!'];
+
+        $this->context
+            ->expects($this->exactly(2))
+            ->method('make')
+            ->willReturnCallback(fn (object $in) => match ($in) {
+                $this->message => $context1,
+                $expected => $context2,
+                default => $this->fail('Unexpected object to convert to context.'),
+            });
+
+
+        $middleware = new LogMessageDispatch($this->logger, LogLevel::NOTICE, LogLevel::WARNING, $this->context);
         $actual = $middleware($this->message, function (Message $received) use ($expected) {
             $this->assertSame($this->message, $received);
             return $expected;
@@ -106,8 +142,8 @@ class LogMessageDispatchTest extends TestCase
 
         $this->assertSame($expected, $actual);
         $this->assertSame([
-            [LogLevel::NOTICE, "Bus dispatching {$name}.", ObjectContext::from($this->message)->context()],
-            [LogLevel::WARNING, "Bus dispatched {$name}.", ResultContext::from($expected)->context()],
+            [LogLevel::NOTICE, "Bus dispatching {$name}.", ['command' => $context1]],
+            [LogLevel::WARNING, "Bus dispatched {$name}.", ['result' => $context2]],
         ], $logs);
     }
 
@@ -120,12 +156,18 @@ class LogMessageDispatchTest extends TestCase
         $message = $this->createMock(Query::class);
         $name = $message::class;
 
+        $this->context
+            ->expects($this->once())
+            ->method('make')
+            ->with($this->identicalTo($message))
+            ->willReturn($context = ['foo' => 'bar']);
+
         $this->logger
             ->expects($this->once())
             ->method('log')
-            ->with(LogLevel::DEBUG, "Bus dispatching {$name}.", ObjectContext::from($message)->context());
+            ->with(LogLevel::DEBUG, "Bus dispatching {$name}.", ['query' => $context]);
 
-        $middleware = new LogMessageDispatch($this->logger);
+        $middleware = new LogMessageDispatch($this->logger, context: $this->context);
 
         try {
             $middleware($message, static function () use ($expected) {